requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
cachelib==0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
Deprecated==1.2.10
feedfinder2==0.0.4
feedparser==6.0.1
howdoi==2.0.7
idna==2.10
jieba3k==0.35.1
joblib==0.17.0
keep==2.9
lxml==4.5.2
newspaper3k==0.2.8
nltk==3.5
Pillow==7.2.0
pyenchant==3.1.1
PyGithub==1.53
Pygments==2.7.1
PyJWT==1.7.1
pyquery==1.4.1
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.10.11
requests==2.24.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
terminaltables==3.1.0
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.50.2
urllib3==1.25.10
wrapt==1.12.1
